+++ freesurfer_surface/__init__.py
@@ -67,7 +67,7 @@ import numpy
 
 try:
     from freesurfer_surface.version import __version__
-except ImportError:  # ModuleNotFoundError not available in python<3.6
+except ModuleNotFoundError:
     # package is not installed
     __version__ = None
 
@@ -161,7 +161,7 @@ class PolygonalCircuit:
 
     def adjacent_vertex_indices(
         self, vertices_num: int = 2
-    ) -> typing.Iterable[typing.Tuple[int]]:
+    ) -> typing.Iterable[typing.Tuple[int, ...]]:
         vertex_indices_cycle = list(
             itertools.islice(
                 itertools.cycle(self.vertex_indices),
@@ -203,9 +203,7 @@ class PolygonalChainsNotOverlapingError(ValueError):
 
 class PolygonalChain:
     def __init__(self, vertex_indices: typing.Iterable[int]):
-        self.vertex_indices = collections.deque(
-            vertex_indices
-        )  # type: typing.Deque[int]
+        self.vertex_indices: typing.Deque[int] = collections.deque(vertex_indices)
 
     def normalized(self) -> "PolygonalChain":
         vertex_indices = list(self.vertex_indices)
@@ -260,12 +258,12 @@ class Label:
     def __init__(
         self, index: int, name: str, red: int, green: int, blue: int, transparency: int
     ):
-        self.index = index  # type: int
-        self.name = name  # type: str
-        self.red = red  # type: int
-        self.green = green  # type: int
-        self.blue = blue  # type: int
-        self.transparency = transparency  # type: int
+        self.index: int = index
+        self.name: str = name
+        self.red: int = red
+        self.green: int = green
+        self.blue: int = blue
+        self.transparency: int = transparency
 
     @property
     def color_code(self) -> int:
@@ -297,9 +295,9 @@ class Annotation:
     _TAG_OLD_COLORTABLE = b"\0\0\0\x01"
 
     def __init__(self):
-        self.vertex_label_index = {}  # type: Dict[int, int]
-        self.colortable_path = None  # type: Optional[bytes]
-        self.labels = {}  # type: Dict[int, Label]
+        self.vertex_label_index: typing.Dict[int, int] = {}
+        self.colortable_path: typing.Optional[bytes] = None
+        self.labels: typing.Dict[int, Label] = {}
 
     @staticmethod
     def _read_label(stream: typing.BinaryIO) -> Label:
@@ -365,14 +363,14 @@ class Surface:
     _DATETIME_FORMAT = "%a %b %d %H:%M:%S %Y"
 
     def __init__(self):
-        self.creator = None  # type: Optional[bytes]
-        self.creation_datetime = None  # type: Optional[datetime.datetime]
-        self.vertices = []  # type: List[Vertex]
-        self.triangles = []  # type: List[Triangle]
-        self.using_old_real_ras = False  # type: bool
-        self.volume_geometry_info = None  # type: Optional[Tuple[bytes]]
-        self.command_lines = []  # type: List[bytes]
-        self.annotation = None  # type: Optional[Annotation]
+        self.creator: bytes = b"pypi.org/project/freesurfer-surface/"
+        self.creation_datetime: typing.Optional[datetime.datetime] = None
+        self.vertices: typing.List[Vertex] = []
+        self.triangles: typing.List[Triangle] = []
+        self.using_old_real_ras: bool = False
+        self.volume_geometry_info: typing.Optional[typing.Tuple[bytes, ...]] = None
+        self.command_lines: typing.List[bytes] = []
+        self.annotation: typing.Optional[Annotation] = None
 
     @classmethod
     def _read_cmdlines(cls, stream: typing.BinaryIO) -> typing.Iterator[bytes]:
@@ -469,6 +467,10 @@ class Surface:
                 self._TAG_OLD_USEREALRAS
                 + struct.pack(">I", 1 if self.using_old_real_ras else 0)
             )
+            if not self.volume_geometry_info:
+                raise ValueError(
+                    "Missing geometry information (set attribute `volume_geometry_info`)"
+                )
             surface_file.write(
                 self._TAG_OLD_SURF_GEOM + b"".join(self.volume_geometry_info)
             )
@@ -507,10 +509,10 @@ class Surface:
     def _triangle_count_by_adjacent_vertex_indices(
         self,
     ) -> typing.Dict[int, typing.DefaultDict[int, int]]:
-        counts = {
+        counts: typing.Dict[int, typing.DefaultDict[int, int]] = {
             vertex_index: collections.defaultdict(lambda: 0)
             for vertex_index in range(len(self.vertices))
-        }  # type: typing.Dict[int, typing.DefaultDict[int, int]]
+        }
         for triangle in self.triangles:
             for vertex_index_pair in triangle.adjacent_vertex_indices(2):
                 counts[vertex_index_pair[0]][vertex_index_pair[1]] += 1
@@ -560,6 +562,10 @@ class Surface:
             yield PolygonalCircuit(cycle_indices)
 
     def _get_vertex_label_index(self, vertex_index: int) -> typing.Optional[int]:
+        if not self.annotation:
+            raise RuntimeError(
+                "Missing annotation (call method `load_annotation_file` first)."
+            )
         return self.annotation.vertex_label_index.get(vertex_index, None)
 
     def _find_label_border_segments(self, label: Label) -> typing.Iterator[LineSegment]:
@@ -607,9 +613,9 @@ class Surface:
     def find_label_border_polygonal_chains(
         self, label: Label
     ) -> typing.Iterator[PolygonalChain]:
-        neighbour_indices = collections.defaultdict(
-            set
-        )  # type: typing.DefaultDict[_VertexSubindex, typing.Set[_VertexSubindex]] # type: ignore
+        neighbour_indices: (  # type: ignore
+            typing.DefaultDict[self._VertexSubindex, typing.Set[self._VertexSubindex]]
+        ) = collections.defaultdict(set)
         for segment in self._find_label_border_segments(label):
             vertex_indices = [(i, 0) for i in segment.vertex_indices]
             neighbour_indices[vertex_indices[0]].add(vertex_indices[1])